RaspiVid: Change --raw-format argument behavior
 - Unknown format produces an error instead of defaulting to YUV
 - --raw is no longer required

Checking raw_output when processing --raw-format causes a "chicken-egg
problem", because "--raw - --raw-format RGB" is valid and "--raw-format
RGB --raw" is invalid. The new approach is consistent with the original
code (i.e. --wrap makes only sense in segment mode but there seems to
be no error message if --wrap is specified outside the segment mode).